lapi/fs: Replace loff_t with long long
loff_t in tst_max_lfs_filesize() would require define _GNU_SOURCE.
Replace with long long to avoid this.

This fixes build error on musl (alpine):

    In file included from unlink09.c:18:
    ../../../../include/lapi/fs.h:58:15: error: unknown type name 'loff_t'
       58 | static inline loff_t tst_max_lfs_filesize(void)

Link: https://lore.kernel.org/ltp/20240425155533.295195-2-pvorel@suse.cz/
Fixes: 2cf78f47a ("unlink: Add error tests for EPERM and EROFS")
Suggested-by: Jan Stancek <jstancek@redhat.com>
Acked-by: Jan Stancek <jstancek@redhat.com>

+++ include/lapi/fs.h
@@ -55,13 +55,13 @@
  * 64 bit: macro taken from kernel from include/linux/fs.h
  * 32 bit: own implementation
  */
-static inline loff_t tst_max_lfs_filesize(void)
+static inline long long tst_max_lfs_filesize(void)
 {
 #ifdef TST_ABI64
-	return (loff_t)LLONG_MAX;
+	return LLONG_MAX;
 #else
         long page_size = getpagesize();
-        loff_t ret = ULONG_MAX;
+        long long ret = ULONG_MAX;
 
         while (page_size >>= 1)
                 ret <<= 1;
